+++ src/Microsoft.SqlTools.ServiceLayer/QueryStore/Contracts/GetHighVariationQueriesReport.cs
@@ -3,7 +3,7 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 //
 
-using Microsoft.SqlTools.SqlCore.QueryDataStore.HighVariation;
+using Microsoft.SqlTools.SqlCore.Performance.HighVariation;
 using Microsoft.SqlTools.Hosting.Protocol.Contracts;
 
 #nullable disable
