+++ src/Microsoft.SqlTools.ServiceLayer/QueryStore/Contracts/GetRegressedQueriesReport.cs
@@ -3,7 +3,7 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 //
 
-using Microsoft.SqlTools.SqlCore.QueryDataStore.RegressedQueries;
+using Microsoft.SqlTools.SqlCore.Performance.RegressedQueries;
 using Microsoft.SqlTools.Hosting.Protocol.Contracts;
 
 #nullable disable
