+++ src/Microsoft.SqlTools.ServiceLayer/QueryStore/QueryStoreService.cs
@@ -6,12 +6,12 @@
 using System;
 using System.Threading.Tasks;
 using Microsoft.Data.SqlClient;
-using Microsoft.SqlTools.SqlCore.QueryDataStore.ForcedPlanQueries;
-using Microsoft.SqlTools.SqlCore.QueryDataStore.HighVariation;
-using Microsoft.SqlTools.SqlCore.QueryDataStore.OverallResourceConsumption;
-using Microsoft.SqlTools.SqlCore.QueryDataStore.PlanSummary;
-using Microsoft.SqlTools.SqlCore.QueryDataStore.RegressedQueries;
-using Microsoft.SqlTools.SqlCore.QueryDataStore.TopResourceConsumers;
+using Microsoft.SqlTools.SqlCore.Performance.ForcedPlanQueries;
+using Microsoft.SqlTools.SqlCore.Performance.HighVariation;
+using Microsoft.SqlTools.SqlCore.Performance.OverallResourceConsumption;
+using Microsoft.SqlTools.SqlCore.Performance.PlanSummary;
+using Microsoft.SqlTools.SqlCore.Performance.RegressedQueries;
+using Microsoft.SqlTools.SqlCore.Performance.TopResourceConsumers;
 using Microsoft.SqlTools.Hosting.Protocol;
 using Microsoft.SqlTools.ServiceLayer.Connection;
 using Microsoft.SqlTools.ServiceLayer.Hosting;
