+++ src/Microsoft.SqlTools.SqlCore/Performance/Common/CustomAttributes.cs
@@ -5,7 +5,7 @@
 
 using System;
 
-namespace Microsoft.SqlTools.SqlCore.QueryDataStore.Common
+namespace Microsoft.SqlTools.SqlCore.Performance.Common
 {
     /// <summary>
     /// Attribute to allow specifying a LocalizedString value for an enum
