+++ src/Microsoft.SqlTools.SqlCore/Performance/Common/QdsMetadataMapper.cs
@@ -10,7 +10,7 @@ using System.Diagnostics;
 using System.Linq;
 using Microsoft.Data.SqlClient;
 
-namespace Microsoft.SqlTools.SqlCore.QueryDataStore.Common
+namespace Microsoft.SqlTools.SqlCore.Performance.Common
 {
     /// <summary>
     /// Class contains logic to map runtime/wait metrics names that are recorded by QDS to equivalent enumerations.
