+++ src/Microsoft.SqlTools.SqlCore/Performance/Common/SelectedTextCoordinates.cs
@@ -3,7 +3,7 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 //
 
-namespace Microsoft.SqlTools.SqlCore.QueryDataStore.Common
+namespace Microsoft.SqlTools.SqlCore.Performance.Common
 {
     /// <summary>
     /// This is a util class which will store starting and ending cordinates of queryText under parent object
