+++ src/Microsoft.SqlTools.SqlCore/Performance/Common/Statistic.cs
@@ -8,7 +8,7 @@ using System.ComponentModel;
 using System.Globalization;
 using System.Linq;
 
-namespace Microsoft.SqlTools.SqlCore.QueryDataStore.Common
+namespace Microsoft.SqlTools.SqlCore.Performance.Common
 {
     /// <summary>
     /// This is a list of aggregations that are supported in query store UI. We call them Statistics.
