+++ src/Microsoft.SqlTools.SqlCore/Performance/Common/TimeInterval.cs
@@ -9,7 +9,7 @@ using System.ComponentModel;
 using System.Globalization;
 using System.Linq;
 
-namespace Microsoft.SqlTools.SqlCore.QueryDataStore.Common
+namespace Microsoft.SqlTools.SqlCore.Performance.Common
 {
     [TypeConverter(typeof(TimeIntervalValueConverter))]
     public enum TimeIntervalOptions
