+++ src/Microsoft.SqlTools.SqlCore/Performance/Controls/ResultSetMapping.cs
@@ -3,7 +3,7 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 //
 
-namespace Microsoft.SqlTools.SqlCore.QueryDataStore.Controls
+namespace Microsoft.SqlTools.SqlCore.Performance.Controls
 {
     public class ResultSetMapping
     {
