+++ src/Microsoft.SqlTools.SqlCore/Performance/HighVariation/HighVariationQueryGenerator.cs
@@ -7,10 +7,10 @@ using System.Collections.Generic;
 using System.Globalization;
 using System.Linq;
 using System.Text;
-using Microsoft.SqlTools.SqlCore.QueryDataStore.Common;
-using Microsoft.SqlTools.SqlCore.QueryDataStore.WaitStats;
+using Microsoft.SqlTools.SqlCore.Performance.Common;
+using Microsoft.SqlTools.SqlCore.Performance.WaitStats;
 
-namespace Microsoft.SqlTools.SqlCore.QueryDataStore.HighVariation
+namespace Microsoft.SqlTools.SqlCore.Performance.HighVariation
 {
     /// <summary>
     /// Class used to generate the queries required for query store UI
