+++ src/Microsoft.SqlTools.SqlCore/Performance/PlanSummary/PlanSummaryQueryGenerator.cs
@@ -5,10 +5,10 @@
 
 using System.Collections.Generic;
 using System.Globalization;
-using Microsoft.SqlTools.SqlCore.QueryDataStore.Common;
-using Microsoft.SqlTools.SqlCore.QueryDataStore.WaitStats;
+using Microsoft.SqlTools.SqlCore.Performance.Common;
+using Microsoft.SqlTools.SqlCore.Performance.WaitStats;
 
-namespace Microsoft.SqlTools.SqlCore.QueryDataStore.PlanSummary
+namespace Microsoft.SqlTools.SqlCore.Performance.PlanSummary
 {
     /// <summary>
     /// Class used to generate the queries required for query store UI
