+++ src/Microsoft.SqlTools.SqlCore/Performance/RegressedQueries/RegressedQueriesQueryGenerator.cs
@@ -8,10 +8,10 @@ using System.Collections.Generic;
 using System.Globalization;
 using System.Linq;
 using System.Text;
-using Microsoft.SqlTools.SqlCore.QueryDataStore.Common;
-using Microsoft.SqlTools.SqlCore.QueryDataStore.WaitStats;
+using Microsoft.SqlTools.SqlCore.Performance.Common;
+using Microsoft.SqlTools.SqlCore.Performance.WaitStats;
 
-namespace Microsoft.SqlTools.SqlCore.QueryDataStore.RegressedQueries
+namespace Microsoft.SqlTools.SqlCore.Performance.RegressedQueries
 {
     /// <summary>
     /// Class used to generate the queries required for query store UI
